requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
loguru==0.5.3
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
python-dotenv==0.15.0
pyzbar==0.1.8
sqlanydb==1.0.10
uvloop==0.14.0; sys_platform=="linux"
